Move the UI thread switch to View-Model
